+++ domain/coords.go
@@ -5,8 +5,8 @@ import "time"
 type Coordinate struct {
 	ID        string    `json: "id"`
 	Time      time.Time `json:"time"`
-	Latitude  float64   `json:"lat"`
-	Longitude float64   `json:"lon"`
+	Latitude  float32   `json:"lat"`
+	Longitude float32   `json:"lon"`
 }
 
 type CoordinateCollection struct {
