+++ domain/outputdata.go
@@ -6,11 +6,11 @@ type OutputData struct {
 	ID           string    `json:"id"`
 	Year         int       `json:"year"`
 	Time         time.Time `json:"time"`
-	StartLat     float64   `json:"startLat"`
-	StartLon     float64   `json:"startLon"`
-	EndLat       float64   `json:"EndLat"`
-	EndLon       float64   `json:"EndLon"`
-	Weight       int       `json:"weight"`
+	StartLat     float32   `json:"startLat"`
+	StartLon     float32   `json:"startLon"`
+	EndLat       float32   `json:"EndLat"`
+	EndLon       float32   `json:"EndLon"`
+	Weight       float32   `json:"weight"`
 	IsOverloaded bool      `json:"isOverloaded"`
 }
