+++ domain/rawinputdata.go
@@ -10,20 +10,20 @@ type RawInputData struct {
 	ID               string    `json:"id"`
 	CollectedTime    time.Time `json:"collectedTime"`
 	InsertedTime     time.Time `json:"insertedTime"`
-	Latitude         float64   `json:"latitude"`
-	Longitude        float64   `json:"longitude"`
-	Weight           int       `json:"weight"`
-	VehicleSpeed     int       `json:"vehicleSpeed"`
-	Acceleration     int       `json:"acceleration"`
-	HeadingDirection int       `json:"headingDirection"`
-	CoolentTemp      int       `json:"coolentTemp"`
-	OilPressure      int       `json:"oilPressure"`
-	IntakeAirTemp    int       `json:"intakeAirTemp"`
-	Rpm              int       `json:"rpm"`
-	EngineLoad       int       `json:"engineLoad"`
-	ElevationAngle   int       `json:"elevationAngle"`
-	O2               int       `json:"o2"`
-	FuelFlow         int       `json:"fuelFlow"`
+	Latitude         float32   `json:"latitude"`
+	Longitude        float32   `json:"longitude"`
+	Weight           float32   `json:"weight"`
+	VehicleSpeed     float32   `json:"vehicleSpeed"`
+	Acceleration     float32   `json:"acceleration"`
+	HeadingDirection float32   `json:"headingDirection"`
+	CoolentTemp      float32   `json:"coolentTemp"`
+	OilPressure      float32   `json:"oilPressure"`
+	IntakeAirTemp    float32   `json:"intakeAirTemp"`
+	Rpm              float32   `json:"rpm"`
+	EngineLoad       float32   `json:"engineLoad"`
+	ElevationAngle   float32   `json:"elevationAngle"`
+	O2               float32   `json:"o2"`
+	FuelFlow         float32   `json:"fuelFlow"`
 }
 
 func (vdata *RawInputData) Bind(r *http.Request) error {
