+++ domain/weatherdata.go
@@ -4,14 +4,14 @@ import "time"
 
 type WeatherData struct {
 	ID                 string    `json:"id"`
-	Latitude           float64   `json:"latitude"`
-	Longitude          float64   `json:"longitude"`
+	Latitude           float32   `json:"latitude"`
+	Longitude          float32   `json:"longitude"`
 	Time               time.Time `json:"time"`
-	WindSpeed          int       `json:"windSpeed"`
-	WindDirection      int       `json:"windDirection"`
-	AtmosphereTemp     int       `json:"atmosphereTemp"`
-	AtmospherePressure int       `json:"atmospherePressure"`
-	Humidity           int       `json:"humidity"`
+	WindSpeed          float32   `json:"windSpeed"`
+	WindDirection      float32   `json:"windDirection"`
+	AtmosphereTemp     float32   `json:"atmosphereTemp"`
+	AtmospherePressure float32   `json:"atmospherePressure"`
+	Humidity           float32   `json:"humidity"`
 }
 
 type WeatherDataCollection struct {
